refactor: extract tag processing utility
